+++ list_people.rb
@@ -1,0 +1,19 @@
+require './app'
+
+class List_People < App
+  def initialize
+    super
+  end
+
+  def list_people
+    if @people_list.empty?
+      p 'There are no people in the list, please create a person7 choosing option 3'
+    else
+      @people_list.each do |person|
+        p "[#{person.class.name}] Name: #{person.name}, ID: #{person.id} Age: #{person.age}"
+      end
+    end
+      puts ''
+      run
+  end
+end
